Use session for Street already Checked

+++ wwwroot/clientSide.js
@@ -3,14 +3,21 @@
     elem.setAttribute("disabled","true");
     }
 
-window.saveInSession = (elementId) => {
-    var val = document.getElementById(elementId).value;
-    var streetsInSession = JSON.parse(window.sessionStorage.getItem('selected'));
+window.saveInSession = (elementId, path) => {
+    //var val = document.getElementById(elementId).value;
+    var streetsInSession = JSON.parse(window.sessionStorage.getItem(path));
     var streets = streetsInSession != null ? streetsInSession : [];
-    streets.push(val);
-    window.sessionStorage.setItem('selected', JSON.stringify(streets));
+    streets.push(elementId);
+    window.sessionStorage.setItem(path, JSON.stringify(streets));
 }
 
-window.sessionStorageClean = () => {
-    window.sessionStorage.clear();
+window.sessionStorageClean = (nrPath) => {
+    var stagedPaths = window.sessionStorage.getItem(nrPath.replace('Nr', ''));
+    if (stagedPaths == null) {
+        window.sessionStorage.clear();
+        return null;
+    }
+    else {
+        return stagedPaths;
+    }
 }
